l04/requirements.txt
ipywidgets==7.5.1
jupyter==1.0.0
matplotlib==3.1.3
numpy==1.17.5
pandas==0.25.3
pyro-ppl==1.3.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
torch==1.4.0

